CG05

+++ src/main/java/com/sisu/sisu/Dao/HistoriaClinicaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.HistoriaClinica;
 
-public interface HistoriaClinicaDao extends CrudRepository<HistoriaClinica, Long>{
+public interface HistoriaClinicaDao extends CrudRepository<HistoriaClinica, Integer>{
     
 }
