+++ src/main/java/com/sisu/sisu/entitys/HistoriaClinica.java
@@ -24,7 +24,7 @@ private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_historia_clinica")
-	private Long idHistoriaClinica;
+	private Integer idHistoriaClinica;
 
 
    @Column(name = "fecha_registro_hc")
